refactor(explain): move the SQLite plan parsers into tabularis-explain
`build_sqlite_tree` takes `(id, parent, detail)` triples and
`parse_sqlite_detail` takes a string — neither touches `sqlx`, so both were
already pure and only their location tied them to the driver. Move them to
`tabularis_explain::sqlite`, leaving the driver to run the statement and
decode rows.

This closes the last gap in the boundary: a host holding SQLite
`EXPLAIN QUERY PLAN` output can now build a plan tree without a driver, as it
already could for Postgres and MySQL/MariaDB. Their three tests move with
them, bringing the crate to 34.
