+++ src-tauri/src/drivers/sqlite/tests.rs
@@ -1,4 +1,3 @@
-use super::explain::{build_sqlite_tree, parse_sqlite_detail};
 use super::sqlite_push_pk_where;
 use super::{alter_view, create_view, drop_view, get_view_columns, get_view_definition, get_views};
 use crate::models::{ConnectionParams, DatabaseSelection};
@@ -45,54 +44,6 @@ async fn setup_test_db() -> (ConnectionParams, NamedTempFile) {
 
     // We return the file handle too so it doesn't get deleted until the test ends
     (params, file)
-}
-
-#[test]
-fn test_parse_sqlite_detail_search_with_primary_key() {
-    let (node_type, relation, index_condition) =
-        parse_sqlite_detail("SEARCH users USING INTEGER PRIMARY KEY (rowid=?)");
-
-    assert_eq!(node_type, "Search");
-    assert_eq!(relation.as_deref(), Some("users"));
-    assert_eq!(index_condition.as_deref(), Some("PRIMARY KEY"));
-}
-
-#[test]
-fn test_parse_sqlite_detail_scan_with_covering_index() {
-    let (node_type, relation, index_condition) =
-        parse_sqlite_detail("SCAN users USING COVERING INDEX idx_users_name");
-
-    assert_eq!(node_type, "Scan");
-    assert_eq!(relation.as_deref(), Some("users"));
-    assert_eq!(index_condition.as_deref(), Some("idx_users_name"));
-}
-
-#[test]
-fn test_build_sqlite_tree_nested_entries() {
-    let entries = vec![
-        (0, 0, "SCAN users".to_string()),
-        (
-            1,
-            0,
-            "SEARCH posts USING INDEX idx_posts_user_id".to_string(),
-        ),
-        (2, 1, "USE TEMP B-TREE FOR ORDER BY".to_string()),
-    ];
-
-    let mut counter = 0;
-    let root = build_sqlite_tree(&entries, 0, &mut counter);
-
-    assert_eq!(root.node_type, "Scan");
-    assert_eq!(root.relation.as_deref(), Some("users"));
-    assert_eq!(root.children.len(), 1);
-    assert_eq!(root.children[0].node_type, "Search");
-    assert_eq!(root.children[0].relation.as_deref(), Some("posts"));
-    assert_eq!(
-        root.children[0].index_condition.as_deref(),
-        Some("idx_posts_user_id")
-    );
-    assert_eq!(root.children[0].children.len(), 1);
-    assert_eq!(root.children[0].children[0].node_type, "Sort");
 }
 
 #[tokio::test]
